XIVY-7371 Better support for selenium driver on a remote host
If selenium driver is remote then the host name of the configured ivy
test engine is replaced if it is localhost, 127.0.0.1 or [::01] to the
real network name of the local host.

This should ensure that the selenium driver still can access the ivy
test engine that is running on local host.